requirements.txt
gym==0.26.1
hydra-core==1.3.2
numpy==1.24.4
omegaconf==2.3.0
Pillow==10.4.0
pytorch_lightning==2.1.2
setproctitle==1.3.3
timm==0.4.12
torch==2.3.1
torchvision==0.18.1
wandb==0.17.4
matplotlib==3.7.2
scikit_learn==1.3.1
tqdm==4.66.1
pillow~=10.3.0
